Trait serde_mtproto::identifiable::Identifiable

pub trait Identifiable {
    fn all_type_ids() -> &'static [u32]
    where
        Self: Sized;

    fn all_enum_variant_names() -> Option<&'static [&'static str]>
    where
        Self: Sized;

    fn type_id(&self) -> u32;

    fn enum_variant_id(&self) -> Option<&'static str>;
}

A trait for a Rust data structure that can have an id.

Required methods

Important traits for &'_ mut [u8]
impl<'_> Write for &'_ mut [u8]impl<'_> Read for &'_ [u8]

fn all_type_ids() -> &'static [u32] where
Self: Sized,

Get all possible ids (known at compile time) of an identifiable type.

This is most useful for enums where each variant has its own id.

Implementation note

This method should return a slice of non-duplicate values, i.e. the slice must effectively be a set with contents known at compile-time. Currently, there is no way to enforce this restriction using the language itself.

This restriction is marked as should because it potentially only alters the behavior of counting all ids a type in question has, but not checking if an arbitrary id is in this set - in this case, a reliable counting routine must traverse the whole slice to eliminate duplicates.

Unfortunately, this worsens the time complexity from O(1) to O(n), but for everyday use-case this is fine since product types we usually use are relatively small to make this a big concern.

Compatibility note

Will probably be replaced by an associated constant after bumping minimum supported Rust version to 1.20.

On the other hand, making static methods associated constants will prevent this trait from usage with dynamic dispatch.

fn all_enum_variant_names() -> Option<&'static [&'static str]> where
Self: Sized,

Get all enum variant names of an identifiable type.

For structs this method must return None and for enums it must return Some with stringified variant names in the same order as the variants themselves.

Compatibility note

Will probably be replaced by an associated constant after bumping minimum supported Rust version to 1.20.

On the other hand, making static methods associated constants will prevent this trait from usage with dynamic dispatch.

fn type_id(&self) -> u32

Get id of a value of an identifiable type.

Its signature is made (&self) -> i32, not () -> i32 because of enum types where different enum variants can have different ids.

Implementation note

This method should return a value contained in the slice returned by all_type_ids() method. Currently, there is no way to enforce this restriction using the language itself.

fn enum_variant_id(&self) -> Option<&'static str>

Get enum variant_hint for a value of an identifiable type.

This method is purely for assisting de::Deserializer to deserialize enum types because Deserialize implementations generated by #[derive(Deserialize)] call Deserializer::deserialize_identifier() to identify an enum variant.

Loading content...