Module serde_mtproto::sized

MtProtoSized trait for any Rust data structure a predictable size of its MTProto binary representation can be computed.

Examples

use serde_mtproto::{MtProtoSized, ByteBuf};

struct Something {
    name: String,
    small_num: u16,
    raw_data: ByteBuf,
    pair: (i8, u64),
}

// Implement manually

impl MtProtoSized for Something {
    fn size_hint(&self) -> serde_mtproto::Result<usize> {
        let mut result = 0;

        result += self.name.size_hint()?;
        result += self.small_num.size_hint()?;
        result += self.raw_data.size_hint()?;
        result += self.pair.size_hint()?;

        Ok(result)
    }
}

let smth = Something {
    name: "John Smith".to_owned(),
    small_num: 2000u16,
    raw_data: ByteBuf::from(vec![0xf4, 0x58, 0x2e, 0x33]),
    pair: (-50, 0xffff_ffff_ffff_ffff),
};

// "John Smith" => 1 byte length, 10 bytes data, 1 byte padding;
// 2000u16 => 4 bytes;
// ByteBuf { ... } => 1 byte length, 4 bytes data, 3 bytes padding;
// (-50, 0xffff_ffff_ffff_ffff) => 4 bytes + 8 bytes == 12 bytes;
//
// Total: 12 + 4 + 8 + 12 == 36 bytes

assert_eq!(36, smth.size_hint()?);

Alternatively, MtProtoSized can be #[derive]d:

#[macro_use]
extern crate serde_mtproto_derive;

#[derive(MtProtoSized)]
struct Something {
    name: String,
    small_num: u16,
    raw_data: Vec<u8>,
    pair: (i8, u64),
}

The derived implementation is the same as the one shown above.

Constants

BOOL_SIZE
Size of a bool MtProto value.

DOUBLE_SIZE
Size of a double MtProto value.

INT128_SIZE
Size of an int128 MtProto value.

INT_SIZE
Size of an int MtProto value.

LONG_SIZE
Size of a long MtProto value.

Traits

MtProtoSized
A trait for a Rust data structure a predictable size of its MTProto binary representation can be computed.

Functions

size_hint_from_byte_seq_len
Helper function for everything naturally representable as a byte sequence.