Trait serde_mtproto::sized::MtProtoSized

pub trait MtProtoSized {
    fn size_hint(&self) -> Result<usize>;
}

A trait for a Rust data structure a predictable size of its MTProto binary representation can be computed.

Required Methods

fn size_hint(&self) -> Result<usize>

Compute the size of MTProto binary representation of this value without actually serializing it.

Returns an error::Result because not any value can be serialized (e.g. strings and sequences that are too long).