Module serde_mtproto::wrappers

Wrapper structs for attaching additional data to a type for [de]serialization purposes.

Data and metadata layout

Wrapper type 	 Layout

Boxed 	 (id, data)

WithSize 	 (size, data)

How does Boxed<WithSize<T>> differ from WithSize<Boxed<T>>?

The first is laid out as (id, size, data) while the second — as (size, id, data). While the id value in both cases represent the type id of data the size value in two layouts above are not the same thing: in the first one it equals data.size_hint()?, but in the second one it equals data.size_hint()? + 4 because it also includes the size of the id value.

Boxed vs WithId

Boxed type has an alias WithId to convey different meanings about it:

Boxed<T> means "not a bare T" where boxed/bare types distinction is drawn from the MTProto official documentation about serialization: https://core.telegram.org/mtproto/serialize#boxed-and-bare-types.

WithId<T> means "T with an id" which explains how this type arranges data and metadata.

This crate uses Boxed as the main naming scheme, whereas WithId is a type alias.

Structs

Boxed
A struct that wraps an Identifiable type value to serialize and deserialize as a boxed MTProto data type.

WithSize
A struct that wraps a MtProtoSized type value to serialize and deserialize as a MTProto data type with the size of its serialized value.

Type Definitions

WithId
An alias for Boxed that is similar to WithSize.