Crate serde_mtproto

Serde MTProto

MTProto is a mobile-first protocol for access to a server API. This crate provides means to serialize Rust types to its binary representation and to deserialize from said representation.

Re-exports

pub use ser::Serializer;

pub use ser::to_bytes;

pub use ser::to_writer;

pub use ser::unsized_bytes_pad_to_bytes;

pub use ser::unsized_bytes_pad_to_writer;

pub use de::Deserializer;

pub use de::from_bytes;

pub use de::from_bytes_reuse;

pub use de::from_reader;

pub use de::from_reader_reuse;

pub use error::Error;

pub use error::ErrorKind;

pub use error::Result;

pub use error::ResultExt;

pub use helpers::UnsizedByteBuf;

pub use helpers::UnsizedByteBufSeed;

pub use helpers::UnsizedBytes;

pub use helpers::size_hint_from_unsized_byte_seq_len;

pub use identifiable::Identifiable;

pub use sized::MtProtoSized;

pub use sized::size_hint_from_byte_seq_len;

pub use wrappers::Boxed;

pub use wrappers::WithId;

pub use wrappers::WithSize;

Modules

de
Deserialize MTProto binary representation to a Rust data structure.

error
When serializing or deserializing MTProto goes wrong.

helpers
Helper types for assisting in some deserialization scenarios.

identifiable
Identifiable trait for any Rust data structure that can have an id.

ser
Serialize a Rust data structure into its MTProto binary representation.

sized
MtProtoSized trait for any Rust data structure a predictable size of its MTProto binary representation can be computed.

wrappers
Wrapper structs for attaching additional data to a type for [de]serialization purposes.

Structs

ByteBuf
Wrapper around Vec<u8> to serialize and deserialize efficiently.

Bytes
Wrapper around &[u8] to serialize and deserialize efficiently.